[MCUX-83023] Fix compilation warnings
A-format, T-format and CANopenNode

+++ digital_encoder_examples/a-format/sync_transfer/a-format_sync_transfer.c
@@ -207,7 +207,7 @@ int main(void)
     }
     else
     {
-        PRINTF("%f (successful)\r\n", temp);
+        PRINTF("%f (successful)\r\n", (double)temp);
     }
     timeout = 0;
